Re-Organise StepSequence Files
