+++ src/components/step-sequence/index.js
@@ -1,0 +1,4 @@
+import StepSequence from './step-sequence.component';
+import StepSequenceItem from './step-sequence-item';
+
+export { StepSequence, StepSequenceItem };
