--- src/components/step-sequence/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "./step-sequence.js"
-}
