+++ src/components/step-sequence/step-sequence-item/index.js
@@ -1,0 +1,1 @@
+export { default } from './step-sequence-item.component';
