--- src/components/step-sequence/step-sequence-item/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "./step-sequence-item.js"
-}
