+++ src/components/step-sequence/step-sequence-item/step-sequence-item-classic.style.js
