+++ src/components/step-sequence/step-sequence-item/step-sequence-item-content.style.js
