+++ src/components/step-sequence/step-sequence-item/step-sequence-item-indicator.style.js
