+++ src/components/step-sequence/step-sequence-item/step-sequence-item-vertical-classic.style.js
