+++ src/components/step-sequence/step-sequence-item/step-sequence-item.style.js
