+++ src/components/step-sequence/step-sequence.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { select } from '@storybook/addon-knobs';
 import StepSequence from './step-sequence.component';
-import StepSequenceItem from './step-sequence-item.component';
+import StepSequenceItem from './step-sequence-item/step-sequence-item.component';
 import OptionsHelper from '../../utils/helpers/options-helper';
 
 storiesOf('Step Sequence', module).add('default', () => {
