--- src/components/step-sequence/style-config.scss
@@ -1,6 +1,0 @@
-@import '../../style-config/colors';
-
-$incomplete-color: $black-55;
-$current-color: $black-85;
-$complete-color: $green;
-$current-link-colour: rgba($complete-color, 0.55);
